Fix: Increase the maximum limit for minSignalBaselineDifference #153
The limit has been increased to 1E+20 and set as double datatype

Issue: #153

+++ libmaven/EIC.cpp
@@ -588,7 +588,7 @@ vector<PeakGroup> EIC::groupPeaks(vector<EIC*>& eics,
                                     double distYWeight,
                                     double overlapWeight,
                                     bool useOverlap,
-                                    float minSignalBaselineDifference) {
+                                    double minSignalBaselineDifference) {
     // Merged to 776
 
     //list filled and return by this function
