+++ libmaven/PeakDetector.h
@@ -141,7 +141,7 @@ public:
 	static vector<EIC*> pullEICs(mzSlice* slice, std::vector<mzSample*>&samples,
 			int peakDetect, int smoothingWindow, int smoothingAlgorithm,
 			float amuQ1, float amuQ3, int baselineSmoothingWindow,
-			int baselineDropTopX, float minSignalBaselineDifference, int eicType);
+			int baselineDropTopX, double minSignalBaselineDifference, int eicType);
 
 private:
 
@@ -168,7 +168,7 @@ struct EicLoader {
 	EicLoader(mzSlice* islice, int eic_type, PeakDetectionFlag iflag = NoPeakDetection,
 			int smoothingWindow = 5, int smoothingAlgorithm = 0, float amuQ1 =
 					0.1, float amuQ2 = 0.5, int baselineSmoothingWindow = 5,
-			int baselineDropTopX = 40, float minSignalBaselineDiff = 0) {
+			int baselineDropTopX = 40, double minSignalBaselineDiff = 0) {
 
 		slice = islice;
 		eicType = eic_type;
@@ -227,7 +227,7 @@ struct EicLoader {
 	float eic_amuQ2;
 	int eic_baselne_smoothingWindow;
 	int eic_baselne_dropTopX;
-	float minSignalBaselineDifference;
+	double minSignalBaselineDifference;
 };
 
 #endif // PEAKDETECTOR_H
