+++ libmaven/eiclogic.cpp
@@ -68,7 +68,7 @@ void EICLogic::groupPeaks(float eic_smoothingWindow,
 							double distYWeight,
 							double overlapWeight,
 							bool useOverlap,
-							float minSignalBaselineDifference) {
+							double minSignalBaselineDifference) {
 
 	peakgroups = EIC::groupPeaks(eics,
 								eic_smoothingWindow,
@@ -97,7 +97,7 @@ mzSlice EICLogic::setMzSlice(float mz, double ppm) {
 void EICLogic::getEIC(mzSlice bounds, vector<mzSample*> samples,
 		int eic_smoothingWindow, int eic_smoothingAlgorithm, float amuQ1,
 		float amuQ3, int baseline_smoothing, int baseline_quantile,
-		float minSignalBaselineDifference, int eicType) {
+		double minSignalBaselineDifference, int eicType) {
 
 	mzSlice slice = _slice;
 	slice.rtmin = bounds.rtmin;
